# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blood Strike Scrim Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a new team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamBody"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamBody"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /teams/{id}/members:
    get:
      operationId: getTeamMembers
      tags: [teams]
      summary: Get team members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: inviteTeamMember
      tags: [teams]
      summary: Invite a user to team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberBody"
      responses:
        "201":
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /teams/{id}/members/{userId}/accept:
    post:
      operationId: acceptTeamInvite
      tags: [teams]
      summary: Accept team invite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invite accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /teams/{id}/members/{userId}:
    delete:
      operationId: removeTeamMember
      tags: [teams]
      summary: Remove team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /scrims:
    get:
      operationId: listScrims
      tags: [scrims]
      summary: List all scrims
      responses:
        "200":
          description: List of scrims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scrim"
    post:
      operationId: createScrim
      tags: [scrims]
      summary: Create a new scrim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScrimBody"
      responses:
        "201":
          description: Created scrim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"
  /scrims/{id}:
    get:
      operationId: getScrim
      tags: [scrims]
      summary: Get scrim by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scrim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"
        "404":
          description: Scrim not found
    patch:
      operationId: updateScrim
      tags: [scrims]
      summary: Update scrim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScrimBody"
      responses:
        "200":
          description: Updated scrim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"
    delete:
      operationId: deleteScrim
      tags: [scrims]
      summary: Delete scrim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /scrims/{id}/register:
    post:
      operationId: registerTeamToScrim
      tags: [scrims]
      summary: Register a team to a scrim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTeamBody"
      responses:
        "201":
          description: Team registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrimRegistration"
  /scrims/{id}/registrations:
    get:
      operationId: getScrimRegistrations
      tags: [scrims]
      summary: Get scrim registrations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScrimRegistration"
...
components:
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        captainId:
          type: integer
        captainUsername:
          type: ["string", "null"]
        totalPoints:
          type: integer
        totalKills:
          type: integer
        totalWins:
          type: integer
        memberCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - logoUrl
        - description
        - captainId
        - captainUsername
        - totalPoints
        - totalKills
        - totalWins
        - memberCount
        - createdAt
    CreateTeamBody:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
        description:
          type: string
        captainId:
          type: integer
      required:
        - name
        - captainId
    UpdateTeamBody:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
        description:
          type: string
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        userId:
          type: integer
        username:
          type: ["string", "null"]
        bloodStrikeId:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected]
        joinedAt:
          type: string
      required:
        - id
        - teamId
        - userId
        - username
        - bloodStrikeId
        - status
        - joinedAt
    InviteMemberBody:
      type: object
      properties:
        userId:
          type: integer
      required:
        - userId
    Scrim:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        scheduledAt:
          type: string
        maxTeams:
          type: integer
        bracketType:
          type: string
          enum: [low, high]
        status:
          type: string
          enum: [open, ongoing, finished]
        maps:
          type: array
          items:
            type: string
        totalRounds:
          type: integer
        flyTimeSeconds:
          type: integer
        weaponBans:
          type: array
          items:
            type: string
        itemBans:
          type: array
          items:
            type: string
        rules:
          type: ["string", "null"]
        createdByUsername:
          type: ["string", "null"]
        registeredTeams:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - scheduledAt
        - maxTeams
        - bracketType
        - status
        - maps
        - totalRounds
        - flyTimeSeconds
        - weaponBans
        - itemBans
        - rules
        - createdByUsername
        - registeredTeams
        - createdAt
    CreateScrimBody:
      type: object
      properties:
        name:
          type: string
        scheduledAt:
          type: string
        maxTeams:
          type: integer
        bracketType:
          type: string
          enum: [low, high]
        maps:
          type: array
          items:
            type: string
        totalRounds:
          type: integer
        flyTimeSeconds:
          type: integer
        weaponBans:
          type: array
          items:
            type: string
        itemBans:
          type: array
          items:
            type: string
        rules:
          type: string
        createdBy:
          type: integer
      required:
        - name
        - scheduledAt
        - maxTeams
        - bracketType
        - maps
        - totalRounds
        - flyTimeSeconds
        - createdBy
    UpdateScrimBody:
      type: object
      properties:
        name:
          type: string
        scheduledAt:
          type: string
        maxTeams:
          type: integer
        bracketType:
          type: string
          enum: [low, high]
        status:
          type: string
          enum: [open, ongoing, finished]
        maps:
          type: array
          items:
            type: string
        totalRounds:
          type: integer
        flyTimeSeconds:
          type: integer
        weaponBans:
          type: array
          items:
            type: string
        itemBans:
          type: array
          items:
            type: string
        rules:
          type: string
    ScrimRegistration:
      type: object
      properties:
        id:
          type: integer
        scrimId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: ["string", "null"]
        registeredAt:
          type: string
      required:
        - id
        - scrimId
        - teamId
        - teamName
        - registeredAt
    RegisterTeamBody:
      type: object
      properties:
        teamId:
          type: integer
      required:
        - teamId
...