# lib/api-zod/src/generated/types/scrimRegistration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Strike Scrim Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ScrimRegistration {
  id: number;
  scrimId: number;
  teamId: number;
  /** @nullable */
  teamName: string | null;
  registeredAt: string;
}